Store ModuleInfo -> ModuleDescriptor mapping inside a SLRU cache
We store this instead of ModuleInfo -> ResolverForModule, so we need to store resolvers somewhere else,
and we use modules 'capabilities' property for sake of it.

The change is necessary because we want to preserve '1 to 1' relation between
resolvers and modules instances.

Otherwise unexpected situations may happen:
- resolver is removed from the cache
- someone has a reference to module descriptor (having a ResolveSession instance indirectly inside)
- than someone requests resolver for module
- new one is created with different ResolveSession instance,
but with same module descriptor instance

So we have two different resolve sessions for one module
and relevant exceptions happening due to this kind of paradox

 #EA-82372 Fixed